The robot can now follow a line pretty well
Need to add more states to the FSM I think.  ITs very jerky following
a line thats layed down in a circle.  When it goes a little off to the
left or the right it seems to still think that its in the center.

Still need to think about how we can identify junctinos for the maze.
Right now the robot would just think that its seeing a line in the
middle and behave like its in the center state.