Added special logic for when the line was lost
If we lost the line, just asume the same reading that
we had previous to losing the ling
Can now make it all the way around the track!
The robot still acts kinda like is drunk, as far as
its swerving all of the time.  I wonder if adding in
the derivative would remove that.. Test for another day